updated makefile, readme, minor refactor for cli deploy
