+++ README.md
@@ -125,6 +125,43 @@ Usage of ./ghost-benchmark:
         Data size of SET/GET value in bytes (default 2)
 ```
 
+### CLI
+Now you can use a simple cli to test or play with data. All the current commands
+are supported. Cli works only if ghost-server exists on provided address.
+
+Run cli:
+```sh
+ghost-cli -host localhost -port 6869
+```
+
+#### Example session:
+```sh
+> ping # will test the connection
+Pong!
+> set hello world # will set value "world" to key "hello"
+OK
+> get hello # will get the value stored with key "hello"
+world
+> del hello # will delete the value stored with key "hello"
+OK
+> cadd mars # will add new "mars" collection
+OK
+> cget mars # will select "mars" collection
+OK
+> set "few words key" "few words value" # if few words in value or keys is needed surround it with quotes
+OK
+> get "few words key"
+few words value
+> set song "riders on the storm" # only one argument could be in quotes if needed
+OK
+> get song
+riders on the storm
+> set "another song" stairway # other order is possible
+OK
+> get "another song"
+stairway
+```
+
 ### Commands
 
 Server commands:
@@ -165,7 +202,7 @@ func main() {
 ```
 
 ## TODO
-  * Implement CLI
+  * Improve CLI
   * Improve server and get rid of limitations
   * Improve documentation
   * Properly comment sources
