+++ cli/user_input_test.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import (
 	"testing"
